+++ groups/bsl/bslstl/bslstl_set.t.cpp
@@ -4444,7 +4444,7 @@ void TestDriver<KEY, COMP, ALLOC>::testCase7_1()
             TestValues VALUES(SPEC);
 
             if (verbose) {
-                printf("\nFor an object of length %d:\n", LENGTH);
+                printf("\nFor an object of length %d:\n", (int)LENGTH);
                 P(SPEC);
             }
 
@@ -4564,7 +4564,7 @@ void TestDriver<KEY, COMP, ALLOC>::testCase7()
             const size_t      LENGTH = (int) strlen(SPEC);
 
             if (verbose) {
-                printf("\nFor an object of length %d:\n", LENGTH);
+                printf("\nFor an object of length %d:\n", (int)LENGTH);
                 P(SPEC);
             }
 
@@ -5206,7 +5206,7 @@ void TestDriver<KEY, COMP, ALLOC>::testCase3()
             Obj mX(&oa);
 
             if ((int)LENGTH != oldLen) {
-                if (verbose) printf("\tof length %d:\n", LENGTH);
+                if (verbose) printf("\tof length %d:\n", (int)LENGTH);
                  ASSERTV(LINE, oldLen <= (int)LENGTH);  // non-decreasing
                 oldLen = LENGTH;
             }
@@ -5397,7 +5397,8 @@ void TestDriver<KEY, COMP, ALLOC>::testCase2()
 
             if (0 < LENGTH) {
                 if (verbose) {
-                    printf("\t\tOn an object of initial length %d.\n", LENGTH);
+                    printf("\t\tOn an object of initial length %d.\n",
+                           (int)LENGTH);
                 }
 
                 for (size_t tj = 0; tj < LENGTH - 1; ++tj) {
